+++ common/src/main/java/net/spacerulerwill/skygrid_reloaded/worldgen/SkyGridConfig.java
@@ -2,22 +2,40 @@ package net.spacerulerwill.skygrid_reloaded.worldgen;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.dimension.LevelStem;
 
-public record SkyGridConfig(SkyGridChunkGeneratorConfig overworldConfig, SkyGridChunkGeneratorConfig netherConfig,
-                            SkyGridChunkGeneratorConfig endConfig) {
-    public static final Codec<SkyGridConfig> CODEC = RecordCodecBuilder.create(
+import java.util.HashMap;
+import java.util.Map;
+
+public class SkyGridConfig {
+    public final Map<ResourceKey<LevelStem>, SkyGridChunkGeneratorConfig> dimensions;
+
+    public SkyGridConfig(Map<ResourceKey<LevelStem>, SkyGridChunkGeneratorConfig> dimensions) {
+        this.dimensions = new HashMap<>(dimensions);
+    }
+
+    public static final Codec<SkyGridConfig> CODEC_V1 = RecordCodecBuilder.create(
             instance -> instance.group(
-                    SkyGridChunkGeneratorConfig.CODEC.fieldOf("overworld_config").forGetter(SkyGridConfig::overworldConfig),
-                    SkyGridChunkGeneratorConfig.CODEC.fieldOf("nether_config").forGetter(SkyGridConfig::netherConfig),
-                    SkyGridChunkGeneratorConfig.CODEC.fieldOf("end_config").forGetter(SkyGridConfig::endConfig)
+                    SkyGridChunkGeneratorConfig.CODEC.fieldOf("overworld_config").forGetter(cfg -> cfg.dimensions.get(LevelStem.OVERWORLD)),
+                    SkyGridChunkGeneratorConfig.CODEC.fieldOf("nether_config").forGetter(cfg -> cfg.dimensions.get(LevelStem.NETHER)),
+                    SkyGridChunkGeneratorConfig.CODEC.fieldOf("end_config").forGetter(cfg -> cfg.dimensions.get(LevelStem.END))
+            ).apply(instance, (overworld, nether, end) -> new SkyGridConfig(
+                    Map.of(LevelStem.OVERWORLD, overworld, LevelStem.NETHER, nether, LevelStem.END, end)
+            ))
+    );
+
+    public static final Codec<SkyGridConfig> CODEC_V2 = RecordCodecBuilder.create(
+            instance -> instance.group(
+                    Codec.unboundedMap(ResourceKey.codec(Registries.LEVEL_STEM), SkyGridChunkGeneratorConfig.CODEC).fieldOf("dimensions").forGetter(skyGridConfig -> skyGridConfig.dimensions)
             ).apply(instance, SkyGridConfig::new)
     );
 
     public SkyGridConfig(SkyGridConfig other) {
         this(
-                new SkyGridChunkGeneratorConfig(other.overworldConfig),
-                new SkyGridChunkGeneratorConfig(other.netherConfig),
-                new SkyGridChunkGeneratorConfig(other.endConfig)
+                Map.copyOf(other.dimensions)
         );
     }
 }